feat(download): Use circuit-breaker and improve retry in sentry
This improves reliability of the sentry downloader by including the
streaming of the body in the retry logic.  This was already done for
the list_files call which retried streaming the json body.

This also introduces a circuit breaker to avoid hammering on sentry if
it had problems during outages.  It takes a shortcut by only having
one global circuit-breaker instead of keeping one per source, because
we know all the sentry-sources will be the same.